shift the database code to models/userModel.js file to increate the readability
