+++ day_6/package.json
@@ -9,6 +9,7 @@
   "author": "sajda",
   "license": "ISC",
   "dependencies": {
+    "bcrypt": "^5.0.1",
     "email-validator": "^2.0.4",
     "express": "^4.18.1",
     "mongoose": "^6.6.1"
